backend/requirements.txt
Django==3.2.11
djangorestframework==3.13.1
django-cors-headers==3.12.0
